Separated SQLAlchemy and Pydantic models into separate submodules
