+++ faddr/parser.py
@@ -5,12 +5,12 @@ from pathlib import Path
 from ttp import ttp
 
 from faddr import logger
-from faddr.database import DeviceModel
 from faddr.exceptions import (
     FaddrParserConfigFileAbsent,
     FaddrParserConfigFileEmpty,
     FaddrParserUnknownProfile,
 )
+from faddr.schemas import DeviceSchema
 
 
 class Parser:
@@ -96,5 +96,5 @@ class Parser:
         """Parse provided configuration and return structured data."""
         parser = ttp(data=self.config, template=self.template)
         parser.parse()
-        result = DeviceModel.parse_obj(parser.result()[0][0]).dict(exclude_unset=True)
+        result = DeviceSchema.parse_obj(parser.result()[0][0]).dict(exclude_unset=True)
         return result
